Create arena class & move functionality to it
Prior to this change, the snake had to know about the game world in
order to check for collission, which meant that it is very hard to
change the size of the game arena or move it elsewhere on screen. Also,
I removed the const keywords from some getters because I don't think we
realistically need them.

This change creates the arena class which will be responsible for game
logic.